+++ src/_utils/keycodes.ts
@@ -1,0 +1,5 @@
+export const KEYCODES = {
+  ESCAPE: 27,
+}
+
+export default KEYCODES